TorontoBarCubesLayout: One Cube / One Side / One LED -> One Light
This is spectacular ;-)

+++ src/MoonLight/Layers/VirtualLayer.h
@@ -51,7 +51,7 @@ struct PhysMap {
     // 2 bytes struct
     struct {                // condensed rgb
       uint16_t rgb : 14;    // 14 bits (554 RGB)
-      uint8_t mapType : 2;  // 2 bits (4)
+      uint16_t mapType : 2;  // 2 bits (4)
     };  // 16 bits
     uint16_t indexP : 14;        // 16384 one physical light (type==1) index to ledsP array
     uint16_t indexesIndex : 14;  // 16384 multiple physical lights (type==2) index in std::vector<std::vector<nrOfLights_t>> mappingTableIndexes;
